perf(F2): move the public API client out of the route-table module
createApiClient lived in public-api.ts, so the browser app and @bb/sdk pulled
the whole publicApiRoutes table (and its schema imports) into the boot closure
just to get the hc<PublicApiRoutes> type. Move the client into api-client.ts
with a type-only import of PublicApiRoutes, re-export it from the barrel, and
give build-bundled-dts its own stub so the bundled plugin .d.ts stays identical.
Verified: turbo typecheck of server-contract/sdk/app/cli/plugin-sdk, the
server-contract and plugin-sdk suites (bundled-types unchanged), and an app
build where public-api.ts is no longer rendered into any boot chunk (before:
24,462 rendered bytes; boot payload 1,603,612 -> 1,601,241 raw bytes).
