+++ packages/plugin-sdk/scripts/build-bundled-dts.mjs
@@ -41,8 +41,18 @@ import { normalizeBundledDts } from "./normalize-bundled-dts.mjs";
 const here = path.dirname(fileURLToPath(import.meta.url));
 const pkgRoot = path.resolve(here, "..");
 const pkgsDir = path.resolve(pkgRoot, "..");
-const publicApiModule = path.join(pkgsDir, "server-contract/src/public-api.ts");
-const publicApiStub = path.join(here, "public-api-stub.d.ts");
+// Server-contract modules whose real declarations are not portable into a
+// flattened .d.ts, each redirected to a loose stub (the stub headers say why).
+const STUBBED_MODULES = new Map([
+  [
+    path.join(pkgsDir, "server-contract/src/public-api.ts"),
+    path.join(here, "public-api-stub.d.ts"),
+  ],
+  [
+    path.join(pkgsDir, "server-contract/src/api-client.ts"),
+    path.join(here, "api-client-stub.d.ts"),
+  ],
+]);
 const outDir = path.join(pkgRoot, "bundled-types");
 const outputs = {
   "bb-plugin-sdk.d.ts": path.join(pkgRoot, "src/index.ts"),
@@ -119,16 +129,18 @@ function resolveBbSource(id) {
 const inlineWorkspace = {
   name: "inline-bb-workspace",
   resolveId(id, importer) {
-    // Redirect server-contract's non-portable route table to the loose stub,
-    // whether imported by bare specifier or by its own barrel's relative path.
+    // Redirect server-contract's non-portable modules to their loose stubs,
+    // whether imported by bare specifier or by a sibling's relative path.
     if (importer) {
       const asTs = path.resolve(
         path.dirname(importer),
         id.replace(/\.js$/, ".ts"),
       );
-      if (asTs === publicApiModule) return publicApiStub;
+      const stub = STUBBED_MODULES.get(asTs);
+      if (stub) return stub;
     }
-    if (id === publicApiModule) return publicApiStub;
+    const stub = STUBBED_MODULES.get(id);
+    if (stub) return stub;
     return resolveBbSource(id);
   },
 };
