+++ packages/server-contract/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./api-client.js";
 export * from "./api-types.js";
 export * from "./api/thread-tabs.js";
 export * from "./common.js";
