+++ packages/server-contract/src/public-api.ts
@@ -1,5 +1,4 @@
 import type { Hono } from "hono";
-import { hc, type ClientRequestOptions } from "hono/client";
 import type {
   AppTheme,
   AppThemeSelection,
@@ -1444,43 +1443,3 @@ export type PublicApiSchema = ApiSchemaFromRouteDescriptors<
 >;
 
 export type PublicApiRoutes = Hono<{}, PublicApiSchema, "/">;
-
-export type PublicApiFetch = (
-  ...args: Parameters<typeof fetch>
-) => ReturnType<typeof fetch>;
-
-/** Omit the options object to use global fetch; provide it to override fetch. */
-export interface PublicApiClientOptions {
-  fetch: PublicApiFetch;
-}
-
-function toHonoClientOptions(
-  options: PublicApiClientOptions | undefined,
-): ClientRequestOptions | undefined {
-  if (options === undefined) {
-    return undefined;
-  }
-  // Hono types custom fetch as typeof fetch, but only calls the function.
-  return { fetch: options.fetch as typeof fetch };
-}
-
-export function createPublicApiClient(
-  baseUrl: string,
-  options?: PublicApiClientOptions,
-) {
-  return hc<PublicApiRoutes>(`${baseUrl}/api/v1`, toHonoClientOptions(options));
-}
-
-export function createApiClient(
-  baseUrl: string,
-  options?: PublicApiClientOptions,
-) {
-  const apiClient = createPublicApiClient(baseUrl, options);
-  return {
-    api: {
-      v1: apiClient,
-    },
-  };
-}
-
-export type ApiClient = ReturnType<typeof createApiClient>;
